requirements.txt
# AI — use google-genai, NOT the deprecated google-generativeai
google-genai>=1.0.0
 
# Server / networking
python-dotenv>=1.0.0
 
# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
mypy>=1.10.0
 